# Conflicts:
#	src/main/java/com/zyyglxt/controller/permissionsController/UserRoleController.java
#	src/main/java/com/zyyglxt/dao/RoleDOMapper.java
#	src/main/java/com/zyyglxt/error/EmBusinessError.java
#	src/main/java/com/zyyglxt/service/impl/ResourcesRoleRefServiceImpl.java
#	src/main/java/com/zyyglxt/service/impl/ResourcesServiceImpl.java
#	src/main/java/com/zyyglxt/service/impl/RoleServiceImpl.java
#	src/main/java/com/zyyglxt/service/impl/UserRoleRefServiceImpl.java
#	src/main/java/com/zyyglxt/service/impl/UserServiceImpl.java
#	src/main/resources/mapping/UserDOMapper.xml
#	src/main/resources/mapping/UserRoleRefDOMapper.xml

+++ src/main/java/com/zyyglxt/service/impl/ChineseMedicineServiceImpl.java
@@ -3,12 +3,16 @@ package com.zyyglxt.service.impl;
 import com.zyyglxt.dao.ChineseMedicineDOMapper;
 import com.zyyglxt.dataobject.ChineseMedicineDO;
 import com.zyyglxt.dataobject.ChineseMedicineDOKey;
+import com.zyyglxt.error.BusinessException;
+import com.zyyglxt.error.EmBusinessError;
 import com.zyyglxt.service.IChineseMedicineService;
+import com.zyyglxt.validator.ValidatorImpl;
+import com.zyyglxt.validator.ValidatorResult;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
 import java.util.List;
-import java.util.UUID;
 
 /**
  * @author qjc
@@ -19,14 +23,18 @@ import java.util.UUID;
 public class ChineseMedicineServiceImpl implements IChineseMedicineService {
 
     @Resource
-    ChineseMedicineDOMapper chineseMedicineDOMapper;
-
+    private ChineseMedicineDOMapper chineseMedicineDOMapper;
+    @Autowired
+    private ValidatorImpl validator;
     /*
     新建名老中医
      */
     @Override
     public int addChineseMedicine(ChineseMedicineDO chineseMedicineDO) {
-        chineseMedicineDO.setItemcode(UUID.randomUUID().toString());
+        ValidatorResult result = validator.validate(chineseMedicineDO);
+        if(result.isHasErrors()){
+            throw new BusinessException(result.getErrMsg(), EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
         return chineseMedicineDOMapper.insertSelective(chineseMedicineDO);
     }
 
@@ -35,6 +43,10 @@ public class ChineseMedicineServiceImpl implements IChineseMedicineService {
      */
     @Override
     public int updateChineseMedicine(ChineseMedicineDO chineseMedicineDO) {
+        ValidatorResult result = validator.validate(chineseMedicineDO);
+        if(result.isHasErrors()){
+            throw new BusinessException(result.getErrMsg(), EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
         return chineseMedicineDOMapper.updateByPrimaryKeySelective(chineseMedicineDO);
     }
 
@@ -43,6 +55,10 @@ public class ChineseMedicineServiceImpl implements IChineseMedicineService {
     */
     @Override
     public int deleteChineseMedicine(ChineseMedicineDOKey chineseMedicineDOKey) {
+        ValidatorResult result = validator.validate(chineseMedicineDOKey);
+        if(result.isHasErrors()){
+            throw new BusinessException(result.getErrMsg(), EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
         return chineseMedicineDOMapper.deleteByPrimaryKey(chineseMedicineDOKey);
     }
 
@@ -57,6 +73,9 @@ public class ChineseMedicineServiceImpl implements IChineseMedicineService {
     * */
     @Override
     public List<ChineseMedicineDO> searchChineseMedicine(String keyWord) {
+        if(keyWord.isEmpty()){
+            throw new BusinessException("关键字不能为空", EmBusinessError.PARAMETER_VALIDATION_ERROR);
+        }
         return chineseMedicineDOMapper.searchChineseMedicine(keyWord);
     }
